genai-chatbot/requirements.txt
streamlit
langchain
pinecone-client
sentence-transformers
unstructured[md,pdf,ppt,xl]
python-docx
pypandoc
pytest
python-magic-bin
python-magic
pyyaml
torch
transformers
aiofiles
asyncio